rename student search input component
