+++ resources/js/Pages/Students/components/StudentSearchInput.vue
